Split each OAuth provider into its own type
Split `GitHub` and `GitLab` OAuthAgent into distinct types for better extensibility.
